fix: Ensure server files never touch any class that imports from net.minecraft.client
