--- fabric/src/main/java/net/impleri/playerskills/server/fabric/PlayerSkillsServerFabric.java
@@ -1,11 +1,0 @@
-package net.impleri.playerskills.server.fabric;
-
-import net.fabricmc.api.DedicatedServerModInitializer;
-import net.impleri.playerskills.server.PlayerSkillsServer;
-
-public class PlayerSkillsServerFabric implements DedicatedServerModInitializer {
-    @Override
-    public void onInitializeServer() {
-        PlayerSkillsServer.init();
-    }
-}
